backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
Flask-SQLAlchemy>=2.5.1
mysql-connector-python>=8.0.20